+++ kamehouse-shell/bin/common/git-functions.sh
@@ -104,23 +104,14 @@ gitCdCommitAllChangesAndPush() {
   gitPushWithRetry ${REMOTE} ${BRANCH}
 }
 
-# Display build version and date of the specified repo
-displayBuildVersionAndDate() {
-  local PROJECT_DIR=$1
-  cd ${PROJECT_DIR}
-
+# This function assumes it's running on a kamehouse git repo root
+getKameHouseBuildVersion() {
   local KAMEHOUSE_RELEASE_VERSION=`grep -e "<version>.*1-KAMEHOUSE-SNAPSHOT</version>" pom.xml | awk '{print $1}'`
   KAMEHOUSE_RELEASE_VERSION=`echo ${KAMEHOUSE_RELEASE_VERSION:9:6}`
-
-  local BUILD_VERSION=`git log | head -n 3 | grep commit | cut -c 8-`
-  BUILD_VERSION=${BUILD_VERSION:0:8}
+  local GIT_COMMIT_HASH=`git rev-parse --short HEAD`
+  local BUILD_VERSION="${GIT_COMMIT_HASH}"
   if [ -n "${KAMEHOUSE_RELEASE_VERSION}" ]; then
     BUILD_VERSION=${KAMEHOUSE_RELEASE_VERSION}"-"${BUILD_VERSION}
   fi
-  echo "buildVersion=${BUILD_VERSION}"
-
-  local BUILD_DATE=`git log | head -n 3 | grep Date | cut -c 9-`
-  BUILD_DATE=${BUILD_DATE:0:25}
-  BUILD_DATE=`date -d"$BUILD_DATE" +%Y-%m-%d' '%H:%M:%S`
-  echo "buildDate=${BUILD_DATE}"
+  echo "${BUILD_VERSION}"
 }