+++ kamehouse-shell/bin/kamehouse/build-kamehouse.sh
@@ -71,22 +71,26 @@ buildProject() {
   checkCommandStatus "$?" "An error occurred building kamehouse"
 
   if [[ "${BUILD_ALL_EXTRA_MODULES}" == "true" || "${MODULE}" == "kamehouse-mobile" ]]; then
-    log.info "Building kamehouse-mobile android app"
-    cd kamehouse-mobile
-    if ${DELETE_ALL_MOBILE_OUTPUTS}; then
-      log.debug "cordova clean ; cordova platform remove android ; cordova platform add android"
-      cordova clean
-      cordova platform remove android
-      cordova platform add android
-    fi
-    ${HOME}/programs/kamehouse-shell/bin/kamehouse/kamehouse-mobile-resync-kh-files.sh
-    cp -v -f pom.xml www/
-    echo "${GIT_COMMIT_HASH}" > www/git-commit-hash.txt
-    date +%Y-%m-%d' '%H:%M:%S > www/build-date.txt
-    log.debug "cordova build android"
-    cordova build android
-    checkCommandStatus "$?" "An error occurred building kamehouse-mobile"
+    buildMobile
   fi
+}
+
+buildMobile() {
+  log.info "Building kamehouse-mobile android app"
+  cd kamehouse-mobile
+  if ${DELETE_ALL_MOBILE_OUTPUTS}; then
+    log.debug "cordova clean ; cordova platform remove android ; cordova platform add android"
+    cordova clean
+    cordova platform remove android
+    cordova platform add android
+  fi
+  ${HOME}/programs/kamehouse-shell/bin/kamehouse/kamehouse-mobile-resync-kh-files.sh
+  cp -v -f pom.xml www/
+  echo "${GIT_COMMIT_HASH}" > www/git-commit-hash.txt
+  date +%Y-%m-%d' '%H:%M:%S > www/build-date.txt
+  log.debug "cordova build android"
+  cordova build android
+  checkCommandStatus "$?" "An error occurred building kamehouse-mobile"
 }
 
 parseArguments() {