+++ kamehouse-shell/bin/kamehouse/kamehouse-cmd-version.sh
@@ -14,23 +14,8 @@ if [ "$?" != "0" ]; then
 fi
 
 main() {
-  displayKameHouseCmdVersion
-}
-
-displayKameHouseCmdVersion() {
-  KAMEHOUSE_CMD_JAR_LS=`ls -l ${HOME}/programs/kamehouse-cmd/lib/kamehouse-cmd* | head -n 1`
-  BUILD_VERSION=`echo ${KAMEHOUSE_CMD_JAR_LS} | awk '{print $9}'`
-  BUILD_VERSION=${BUILD_VERSION##*/}
-  BUILD_VERSION=`echo ${BUILD_VERSION} | cut -d'-' -f 3`
-  GIT_COMMIT_HASH=`cat ${HOME}/programs/kamehouse-cmd/lib/git-commit-hash.txt 2>/dev/null`
-  if [ -n "${GIT_COMMIT_HASH}" ]; then
-    BUILD_VERSION=${BUILD_VERSION}"-"${GIT_COMMIT_HASH}
-  fi
-  echo "buildVersion=${BUILD_VERSION}"
-  KAMEHOUSE_CMD_JAR_FILE=`ls ${HOME}/programs/kamehouse-cmd/lib/kamehouse-cmd* | head -n 1`
-  BUILD_DATE=`stat ${KAMEHOUSE_CMD_JAR_FILE} | grep "Modify"` 
-  BUILD_DATE=${BUILD_DATE:8:19}
-  echo "buildDate=${BUILD_DATE}"
+  local CMD_VERSION_FILE="${HOME}/programs/kamehouse-cmd/lib/cmd-version.txt"
+  cat "${CMD_VERSION_FILE}"
 }
 
 main "$@"